refactored AccountInfo tests to single file using pytest
